I've added in the cmake code to get Git information for the python wrappers.
Also, I've found the cause of the link problems of Sire against
libOpenMM.so on Linux. The problem occurs when using a C++ 11 compiler
to link against the libOpenMM.so library that comes with miniconda. This
library was compiled without C++ 11 support, so doesn't have the necessary
symbols for functions called with arguments C++11-std::string.

The workaround for now is to disable C++ 11 support when compiling
on Linux, or to use an older compiler (e.g. doesn't work with
GCC 5.2.0 but does work with GCC 4.9.3

[ci skip]