I've update threading building blocks to the latest version.
I've also updated the compiler flags to switch on C++ 11, 1Y or 14
support if available, and to raise an error if at least one of
these is not supported (I want to start writing in C++ 11, and
every modern compiler is now on C++ 14, so this is reasonable).

This causes a link error with GCC 5.2 against libOpenMM, as libOpenMM
that comes with conda install openmm does not expose any C++ 11
functions (e.g. cannot link to functions taking a std::__cxx11::basic_string).