I've fixed a bug in evaluate().mass() which was calculating the wrong mass for partial molecules. I have added in SIRE??_EXPORT to template classes in an attempt to fix the incorrect C++ template wrapper error for Selector<T> when it is converted from a PropPtr<MoleculeView>, e.g.
print( mol.atoms( AtomName("CA") ) ) works, but
print( mol[ AtomName("CA") ] ) doesn't,

as the "wrong" Selector<T> is returned by the wrapper (not the one
that is instantiated in libSireMol).

As this triggers a full recompile, I have taken the opportunity
of adding "operator-" support to the ID classes, as Python won't
allow me to overload "operator!", and so I am using "-" as the
"not" operator, e.g.

mol[ ResName("ALA") - AtomName("CA") ]

should return all atoms that are not called "CA" in the residues
called "ALA", and

mol[ -ResName("ALA") ]

will return all residues that are not called "ALA"