Fixed the ViewsOfMol integration problems for ligandswap and quantomm.
However, now the long run simulations for waterswap, proteinswap, ligandswap
and quantomm are showing weird behaviour. The free energies are not good
and they are not recording and printing out the residue-based components.

I will debug this by automating the integration tests as part of the
new cibod server in January. Until then, have a merry christmas
and a happy new year

[ci skip]