I've fixed several bugs related to conversion between gromacs and amber formats. The issue was that there were bugs in the nb14 scaling factors (not symmetric) and also that amber needs to have empty dihedrals added for 1-4 pairs that should be scaled, even if the dihedral potential is zero.
Now have complete agreement in energy from systems loaded
to and from gromacs and amber prm format files :-)

Only issue now is that Sire and gromacs disagree about the
dihedral energy (agree on everything else). I think this
is because we are assigning wildcard dihedrals differently...
will investigate this afternoon!

[ci skip]