Found one of the problems - I was not allowing for multiple dihedrals to resolve to an amber-matched dihedral. This meant that I only got part of teh function. I have now allowed multiple, and have got most of the dihedral calcualtion working.
The problem now is that there are some dihedrals that seem to exist
in the top file, and that Sire evaluates as having an energy, but
which gromacs ignores. Gromacs calculates their energy if they
are written out by Sire (gromacs reading the Sire-generated
topology), and then Sire and gromacs agree on the dihedral
energy.

However, noticed that Gromacs doesn't calculate a 1-4 energy
automatically from Sire-generated topologies. I can't see why,
as I have it set to auto-generate! I added code (now disabled)
that manually generates the pairs section (which is generated
to be identical to the original kigaki.top file). However,
grommp fails to parse this as it is missing a default
lj 14 term... Where this comes from I don't know!

This is all for now. Most of gromacs is done and I am now
stuck with small details and tracking down missing energies...

[ci skip]