Cleaned up the GeneralUnit class so that it understands both zero and dimensionless values. This lets it cope better with such values.
Added in `isZero` and `isDimensionless` function to test if a unit
is dimensionless or zero. Added code that automatically converted a unit to zero
if it became that value.

Added in a `apply_reduce` function that can perform reductions on the results
of `apply` on molecule views. This can be used to do sums or comparisons,
e.g. find the maximum residue mass

>>> print(mol.residues().apply("mass")
[43.044 g mol-1, 71.08 g mol-1, 30.052 g mol-1]
>>> print(mol.residues().apply_reduce("mass", max))
71.08 g mol-1
>>> print(mol.residues().apply("apply_reduce", "mass", max))
[16 g mol-1, 16 g mol-1, 14.01 g mol-1]

(this last one finds the mass of the heaviest atom in each residue)

Doing this so that the residue/chain/segment properties chapter
has something interesting to write about