Fixed the bugs that stopped waterswap from working. This was caused by the change in BondID that meant that it was always ordered in order of increasing atomidx number. This isn't what ZmatrixMaker expected, and so it couldn't detect where in the protein chain each residue was positioned. All fixed :-)
This then revealed a crash caused by incorrect calculation of the cutgroups
that change for a move. I have "fixed" this by detecting it and returning
that the whole molecule changed (including printing an error).

This works, but is slower and does need fixing (if only to remove the
warning message!)