Begun work on the tutorial describing how to move atoms and molecules.
First, the simple route, by directly editing the coordinates property.

In doing this, I realised that it would be nice if tuples/lists could
auto-convert to Vectors when they are added to other Vectors.

So this commit also adds that functionality.

You can now write;

>>> import sire as sr
>>> v = sr.maths.Vector(1,2,3)
>>> v + (1,2,3)
( 2 Å, 4 Å, 6 Å )

>>> v - (1,1,1)
( 0 , 1 Å, 2 Å )

>>> v += (1,2,3)
>>> v
( 2 Å, 4 Å, 6 Å )

>>> v -= (1,2,3)
>>> v
( 1 Å, 2 Å, 3 Å )

The units are the current default length units. You can specify them
if you want, e.g.

>>> v + (5*sr.units.nanometer, 0, 0)
( 51 Å, 2 Å, 3 Å )