Needed a further fix to the property search to cope with actual string properties (these were not showing up unless they actually equalled 'True').
The logic is that a `property {name}` search is converted to
a `property {name} == True` search. The logic for `== True`
is that the value exists and does not have a value that
matches `False` in any way (e.g. a zero or any case
string that matches "false"). This seems to work.