Added in the code to set and change internals (e.g. bonds, angles, dihedrals) using a Cursor/Cursors/CursorsM object.
This supports an API such as

>>> cursor = mols.cursor()
>>> cursor.angles("element H", "element O", "element H").set_size(104.5)
>>> mols = cursor.commit()

or

>>> cursor = mol.cursor()
>>> cursor.bonds("element H", "element C").set_length(1.09)
>>> mol = cursor.commit()

etc

I need to fully test this, plus also add on some code that better
exposes the weight function, anchors and post-move alignment
options. I also will likely update the dihedral move so that
it rotates all atoms around the dihedral by default, with
an option to disable this as needed.