I've cleaned up the grammar code and updated the tests where brackets are needed to specify precedence.
It is now working (nearly) as I want. Distance searches now use expressions,
and the entire expression hierarchy now works properly with a
well-defined LHS and RHS grammar.

The only niggle is that the grammar actually ignores spaces...

meaning that `mols["bondswithatomsinresnameALA"]` works...

I will have to fix this as it will lead to confusion when spaces
are needed, and where these strings accidentally clash with
user-provided tokens.